+++ crates/youki/src/commands/state.rs
@@ -1,8 +1,10 @@
-use std::path::PathBuf;
+use std::collections::HashMap;
+use std::path::{Path, PathBuf};
 
 use anyhow::Result;
-use libcontainer::container::state::StateExporter;
+use libcontainer::container::state::{ContainerStatus, State as LibState};
 use liboci_cli::State;
+use serde::Serialize;
 
 use crate::commands::load_container;
 
@@ -11,4 +13,30 @@ pub fn state(args: State, root_path: PathBuf) -> Result<()> {
     let export = StateExporter::from(&container.state);
     println!("{}", serde_json::to_string_pretty(&export)?);
     std::process::exit(0);
+}
+
+#[derive(Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct StateExporter<'a> {
+    pub oci_version: &'a str,
+    pub id: &'a str,
+    pub status: ContainerStatus,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub pid: Option<i32>,
+    pub bundle: &'a Path,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub annotations: Option<&'a HashMap<String, String>>,
+}
+
+impl<'a> From<&'a LibState> for StateExporter<'a> {
+    fn from(state: &'a LibState) -> Self {
+        Self {
+            oci_version: &state.oci_version,
+            id: &state.id,
+            status: state.status,
+            pid: state.pid,
+            bundle: &state.bundle,
+            annotations: state.annotations.as_ref(),
+        }
+    }
 }
